Remove hard-coded assumptions about font widths
And instead base calculations on UILabel.sizeThatFits().

Which paves the way for being able to use proportional fonts in the
future.

Changed DataItemBase sizing to support an explicit "prefix" whose width
is used to indent all subsequent lines, and instead of a format method
which passes in a size in characters, getText passes in a checkFit
function which can be called to see if a particular text string will
fit in a single line or not.