Completely rejigged how image fetch and render is done
And how redacted images are handled. We no longer use a second deviceid
as a workaround for the download getting too big, instead there's a new
image format which consists of an unencrypted header and table of
contents, and multiple RLE-encoded images encrypted separately.
Currently there'll always be two images (the rendered calendar, and the
configured privacy image) but potentially there could be more. This is
streamed directly to the flash rather than buffering the entire thing in
memory, this is no longer possible with poorly-compressing privacy
images given the RAM limitations of the esp32.

The workflow when the privacy image is shown on the panel is now the
same as for the calendar, in that updates are still polled for, so if
the privacy image is reconfigured but the calendar is the same, this
will correctly get rerendered on the next update. Previously there were
all kinds of corner cases where the panel behaved differently if the
privacy image was being shown.

The app is now smarter about not reuploading images that haven't
changed, meaning the panel has to do less work (and hammer the flash
less) before establishing there's no changes. This is important because
we now have to write the entire payload to flash before we can calculate
the hashes. The panel still maintains individual images hashes to
optimise away refreshing the e-ink display.

The way images are rendered onto the panel is now completely different,
after they are decrypted from the payload blob they are pre-parsed into
a form optimised for writing directly out over the SPI bus, and that is
what's saved to flash. This saves about 30 seconds when configuring the
display. As a result the on-disk format is uncompressed and much larger,
necessitating another tweak to the partition table. Fortunately we have
lots of flash to spare.

Finally, panel.lua and network.lua are now proper modules with separate
namespaces. Burned some old esp8266 code that's no longer worth keeping.