fix: Better display of non-all-day events that span multiple days (#72)
To match what native Calendar app does, for an event that starts on day
A, continues through day B (and C, D, etc) and ends on day E, the event
is shown as an all-day event on all the interim days, and on day E only
the end time is shown. Previously the start time on day A would
incorrectly be shown on all the other days too.

As part of this, added support for multiline prefix strings in
ViewController.renderToImage().