ci: Enable signed iOS builds and TestFlight uploads (#80)
This change introduces semantic versioning for determining the iOS app version, but does not make any effort to automatically publish releases as this is inevitably a manaul process with the App Store. It resets the iOS app base version back to 0.1 (instead of 1.0) in anticipation of a number of pre-release builds.