ci: Set up automated firmware builds (#90)
This change introduces a new script (`firmware/build.sh`) that will check out and build @tomsci's branch of the NodeMCU firmware, and then use the resulting tools to build the LFS from the `nodemcu` directory. It uses Docker to perform the actual build and supports incremental builds so should be good for just rebuilding LFS during a regular development cycle. Performance seems pretty good on Linux, but is dire on macOS (hey, at least it works).