ci: Perform smoke test builds on shipping Xcode (#122)
Right now, the Monterey builder is using a beta build of Xcode. Since we're not all using this beta, it's important to ensure the project can still build on Xcode 12. This change enables builds on GitHub's `macos-10.15` runner which (at the time of this change) guarantee's we'll get Xcode 12.

Unfortunately, due to an Xcode bug, archive builds fail on Xcode 12, so the change also adds a flag to disable archive builds on 10.15 builds and relies on just the smoke test builds.