fix(ios): Verticial divider respects the height of the heading font (#161)
This change makes the render state pertaining to the divider more explicit by introducing an `DividerStyle` enum with associated types for tracking both the divider orientation and position. It also introduces a drive-by fix to make the naming of the panel dimensions more idiomatic (dropping the `k` prefix), and pulls the status bar height out into a constant.