refactor(ios): Use `DataItemFlags.prefersEmptyColumn` to express column wrapping preference (#164)
This change goes some way to making the layout more explicit by no longer using the condition that a header is not the first item as an indication that content should wrap to another column. Instead, it makes this explicit by using the `.prefersEmptyColumn` flag. This way, non-header items can also request a new column, and the `.header` flag can simply be used to determine the style. The change also introduces a new layout variable (`columnItemCount`) for tracking the number of items in the current column and uses this to determine whether a column wrap is actually necessary.