refactor(ios): Introduce a `DataItemFlags.spansColumns` flag for full-width items (#165)
This change removes the last bit of special-case code in the layout engine that relies on item index in the list to determine layout. Instead of hard-coding that the first item in the list is full-width if it's a header, it checks for the new `.spansColumns` flag which can be applied to an item at any position in the list.