refactor(ios): Introduce an explicit source for generating date headers (#167)
This change introduces `CalendarHeaderSource` which provides a number of formatting options for generating different date headers. The long-term plan is to also use this to generate a localized verison of the "Tomorrow:" text which is currently displayed in the second column, and add configuration options to allow users to choose how their dates are displayed.