feat(iOS): About screen with app version, licenses and credits (#187)
This change introduces a new about screen (accessed from Settings) which includes a list of contributors, credits, details of the app version, and the licenses of the various fonts used. This is also intended to serve as a location of the in-app privacy policy and company details (which are I, I believe, required for App Store submission). Since the fonts licenses are now available in the about screen, they've been removed from the font pickers.

This change also increases the minimum required iOS version to iOS 14.