fix(ios): Tighten line spacing on unifont-16
Previously it was hard-coded to 10, now use half the line height.
Meaning for unifont-16, this reduces the inter-item gap to 8 pixels,
which I think looks better.

Also made 10px be the maximum so that unifont-32 doesn't look overly
spaced out.

Also made DummyDataSource honour showIcons as oftentimes showing all
the special chars in one big block isn't that useful and showIcons
offers a convenient way of hiding that for testing.