refactor(ios): Encapsulate the remote URL parsing code (#263)
This change moves the remote URL handling from `application(_:open:options:)` to a new `ExternalOperation` struct and uses this to both generate URLs for adding dummy devices and parsing remote URLs from the Camera QR code scanner, and from `QRCodeViewController`. This is preparatory work for reusing `WifiProvisionerController` in a full wizard flow for adding devices.