ci: Use a submodule for firmware builds (#152)
This change adds the nodemcu-firmware repository as a submodule (pointing to the head of @tomsci's branch), and uses this for firmware builds. This makes the firmware version tracking explicit and will hopefully make it much easier for us to manage updates in the future.

It's only a partial change as the firmware we're building still doesn't actually run on device (it turns out we will also need to pass `MORE_CFLAGS="-DLUA_NUMBER_INTEGRAL"` to the `make` commnad), but that fix is planned to land in a future change which moves to a custom `Dockerfile` for building as the pre-built image is overkill for our needs.