fix(ios): Select sensible first-run title and body fonts (#357)
This change selects ChiKareGo 2 as the default title font and Unifont 16pt as the default body font. In order to make it a little safer to refer to fonts by identifier in code, it also introduces some static variables scoped to `Fonts.FontName` for the identifiers. Since these are meant to be unique, we'll probably want to make this an enum in the future.