fix(ios): Image preview flashes when pulling 'Privacy Mode' view down (#366)
The privacy image preview was flashing when pulling the 'Privacy Mode' view controller down a little (as if to dismiss) because it was triggering a reload of the privacy image, even though it had already been loaded. This change guards that with a `firstRun` flag to prevent unnecessary loads and subsequent cross-fade animations (leading to the flash). Long-term (and especially if we ever switch this view to SwiftUI), we'll want to create a little privacy image manager that's responsible for explicitly publishing the loading state and performing the privacy image transform.