fix(device-python): Update the display when network connectivity returns (#474)
The change to guard updates against over updating by storing the
'last-modified' header introduced a bug whereby last modified was never
cleared if the display was updated to show an error. This change clears
the cached last modified field in the case of an error to ensure the
next successful fetch updates the display.