fix(ios): Show a horizontal divider in single-column mode
We were incorrectly inserting a vertical divider whenever we saw the 'spans columns' flag _even_ if we were in single-column mode. This change fixes that bug and renames the 'Prefers New Column' option to 'Prefers New Section' to acknowledge that it will add a divider in both single- and two-column mode.