ci(service): Configure the APNS service for background push notifications
This change introduces some guards in the web service that should ensure it fails to start up if APNS is incorrectly configured. This should serve to ensure the service doesn't pass tests or look like everything is working correctly if we can't send push notifications. The change also injects the APNS configuration into the server as part of the GitHub Actions deployment.